Deleting reporter from TestRunner, DefaultTestReporter was folded into a common implementation for Console and other renderers
