+++ test/shared/src/main/scala/zio/test/ExecutionEventSink.scala
@@ -1,6 +1,6 @@
 package zio.test
 
-import zio.{Ref, UIO, ZIO, ZLayer}
+import zio.{Ref, UIO, ULayer, ZIO, ZLayer}
 
 trait ExecutionEventSink {
   def getSummary: UIO[Summary]
@@ -21,7 +21,7 @@ object ExecutionEventSink {
 
   def ExecutionEventSinkLive(testOutput: TestOutput): ZIO[Any, Nothing, ExecutionEventSink] =
     for {
-      summary <- Ref.make[Summary](Summary(0, 0, 0, ""))
+      summary <- Ref.make[Summary](Summary.empty)
     } yield new ExecutionEventSink {
 
       override def process(
@@ -44,5 +44,14 @@ object ExecutionEventSink {
         testOutput <- ZIO.service[TestOutput]
         sink       <- ExecutionEventSinkLive(testOutput)
       } yield sink
+    )
+
+  val silent: ULayer[ExecutionEventSink] =
+    ZLayer.succeed(
+      new ExecutionEventSink {
+        override def getSummary: UIO[Summary] = ZIO.succeed(Summary.empty)
+
+        override def process(event: ExecutionEvent): ZIO[Any, Nothing, Unit] = ZIO.unit
+      }
     )
 }
