+++ test/shared/src/main/scala/zio/test/ReporterEventRenderer.scala
@@ -1,25 +1,26 @@
 package zio.test
 
-import zio.Chunk
+import zio.stacktracer.TracingImplicits.disableAutoTrace
+import zio.{Chunk, Trace}
 import zio.test.render.{ConsoleRenderer, IntelliJRenderer}
 
 trait ReporterEventRenderer {
-  def render(event: ExecutionEvent): Chunk[String]
+  def render(event: ExecutionEvent)(implicit trace: Trace): Chunk[String]
 }
 object ReporterEventRenderer {
   object ConsoleEventRenderer extends ReporterEventRenderer {
-    override def render(executionEvent: ExecutionEvent): Chunk[String] =
+    override def render(executionEvent: ExecutionEvent)(implicit trace: Trace): Chunk[String] =
       Chunk.fromIterable(
         ConsoleRenderer
-          .render(DefaultTestReporter.render(executionEvent, includeCause = true), TestAnnotationRenderer.timed)
+          .render(executionEvent, includeCause = true)
       )
   }
 
   object IntelliJEventRenderer extends ReporterEventRenderer {
-    override def render(event: ExecutionEvent): Chunk[String] =
+    override def render(executionEvent: ExecutionEvent)(implicit trace: Trace): Chunk[String] =
       Chunk.fromIterable(
         IntelliJRenderer
-          .render(event, includeCause = false)
+          .render(executionEvent, includeCause = false)
       )
   }
 }
