+++ test/shared/src/main/scala/zio/test/TestAnnotationRenderer.scala
@@ -16,7 +16,6 @@
 
 package zio.test
 
-import zio.duration2DurationOps
 import zio.stacktracer.TracingImplicits.disableAutoTrace
 import zio.test.TestAnnotationRenderer._
