+++ test/shared/src/main/scala/zio/test/ZIOSpecAbstract.scala
@@ -18,9 +18,9 @@ package zio.test
 
 import org.portablescala.reflect.annotation.EnableReflectiveInstantiation
 import zio._
-import zio.internal.stacktracer.Tracer
 import zio.stacktracer.TracingImplicits.disableAutoTrace
-import zio.test.render._
+import zio.test.ReporterEventRenderer.{ConsoleEventRenderer, IntelliJEventRenderer}
+import zio.test.render.{ConsoleRenderer, TestRenderer}
 
 @EnableReflectiveInstantiation
 abstract class ZIOSpecAbstract extends ZIOApp with ZIOSpecAbstractVersionSpecific {
@@ -29,12 +29,7 @@ abstract class ZIOSpecAbstract extends ZIOApp with ZIOSpecAbstractVersionSpecifi
   def spec: Spec[Environment with TestEnvironment with ZIOAppArgs with Scope, Any]
 
   def aspects: Chunk[TestAspectAtLeastR[Environment with TestEnvironment with ZIOAppArgs]] =
-    Chunk.empty
-
-  def testReporter(testRenderer: TestRenderer, testAnnotationRenderer: TestAnnotationRenderer)(implicit
-    trace: Trace
-  ): TestReporter =
-    ??? // DefaultTestReporter(testRenderer, testAnnotationRenderer)
+    Chunk(TestAspect.fibers)
 
   final def run: ZIO[ZIOAppArgs with Scope, Any, Summary] = {
     implicit val trace = Trace.empty
@@ -75,21 +70,17 @@ abstract class ZIOSpecAbstract extends ZIOApp with ZIOSpecAbstractVersionSpecifi
       args    <- ZIO.service[ZIOAppArgs]
       console <- ZIO.console
       testArgs = TestArgs.parse(args.getArgs.toArray)
-      _       <- ZIO.debug("runSpec")
       summary <- runSpecInfallible(spec, testArgs, console)
-      _ <- ZIO.when(summary.status == Summary.Failure)(
-             ZIO.fail(new Exception("Failed tests."))
-           )
+      _ <- ZIO.when(testArgs.printSummary) {
+             console.printLine(ConsoleRenderer.renderSummary(summary)).orDie
+           }
     } yield summary
 
-  private def createTestReporter(rendererName: String)(implicit trace: Trace): TestReporter = {
-    val renderer = rendererName match {
-      case "intellij" => IntelliJRenderer
-      case _          => ???
+  private def getTestEventRenderer(testArgs: TestArgs) =
+    testArgs.testRenderer match {
+      case Some("intellij") => IntelliJEventRenderer
+      case _                => ConsoleEventRenderer
     }
-    //testReporter(renderer, TestAnnotationRenderer.default)
-    ???
-  }
 
   /*
    * Regardless of test assertion or runtime failures, this method will always return a summary
@@ -98,89 +89,37 @@ abstract class ZIOSpecAbstract extends ZIOApp with ZIOSpecAbstractVersionSpecifi
   private[zio] def runSpecInfallible(
     spec: Spec[Environment with TestEnvironment with ZIOAppArgs with Scope, Any],
     testArgs: TestArgs,
-    console: Console
+    console: Console,
+    runtime: Option[Runtime[_]] = None,
+    testEventHandler: ZTestEventHandler = ZTestEventHandler.silent
   )(implicit
     trace: Trace
-  ): URIO[
-    TestEnvironment with ZIOAppArgs with Scope,
-    Summary
-  ] = {
+  ): URIO[TestEnvironment with ZIOAppArgs with Scope, Summary] = {
     val filteredSpec = FilteredSpec(spec, testArgs)
 
     for {
-      runtime <-
-        ZIO.runtime[
-          TestEnvironment with ZIOAppArgs with Scope
-        ]
+      runtime <- ZIO.fromOption(runtime.collect { case rt: Runtime[Environment with ZIOAppArgs with Scope] =>
+                   rt
+                 }) orElse ZIO.runtime[TestEnvironment with ZIOAppArgs with Scope]
       environment0: ZEnvironment[ZIOAppArgs with Scope] = runtime.environment
       environment1: ZEnvironment[ZIOAppArgs with Scope] = runtime.environment
       sharedLayer: ZLayer[Any, Any, Environment] =
         ZLayer.succeedEnvironment(environment0) >>> bootstrap
       perTestLayer = (ZLayer.succeedEnvironment(environment1) ++ liveEnvironment) >>> (TestEnvironment.live ++ ZLayer
                        .environment[Scope] ++ ZLayer.environment[ZIOAppArgs])
-      executionEventSinkLayer = sinkLayerWithConsole(console)
+      eventRenderer           = getTestEventRenderer(testArgs)
+      executionEventSinkLayer = sinkLayer(console, eventRenderer)
       runner =
         TestRunner(
           TestExecutor
-            .default[
-              Environment,
-              Any
-            ](
+            .default[Environment, Any](
               sharedLayer,
               perTestLayer,
               executionEventSinkLayer,
-              _ => ZIO.unit
+              testEventHandler
             )
         )
-      testReporter = testArgs.testRenderer.fold(runner.reporter)(createTestReporter)
-      summary <-
-        runner.withReporter(testReporter).run(aspects.foldLeft(filteredSpec)(_ @@ _) @@ TestAspect.fibers)
+      summary <- runner.run(aspects.foldLeft(filteredSpec)(_ @@ _) @@ TestAspect.fibers)
     } yield summary
   }
-
-  private[zio] def runSpecInfallible(
-    spec: Spec[Environment with TestEnvironment with ZIOAppArgs with Scope, Any],
-    testArgs: TestArgs,
-    console: Console,
-    runtime: Runtime[_],
-    eventHandlerZ: ZTestEventHandler
-  )(implicit
-    trace: Trace
-  ): URIO[
-    TestEnvironment with ZIOAppArgs with Scope,
-    Summary
-  ] = {
-    val filteredSpec = FilteredSpec(spec, testArgs)
-
-    val castedRuntime: Runtime[Environment with ZIOAppArgs with Scope with ExecutionEventSink] =
-      runtime.asInstanceOf[Runtime[Environment with ZIOAppArgs with Scope with ExecutionEventSink]]
-
-    for {
-      _                                                <- ZIO.unit
-      environment1: ZEnvironment[ZIOAppArgs with Scope] = castedRuntime.environment
-      sharedLayer: ZLayer[Any, Nothing, Environment with ExecutionEventSink] =
-        ZLayer.succeedEnvironment(castedRuntime.environment)
-      perTestLayer: ZLayer[Any, Nothing, TestEnvironment with Scope with ZIOAppArgs] =
-        (ZLayer.succeedEnvironment(environment1) ++ liveEnvironment) >>> (TestEnvironment.live ++ ZLayer
-          .environment[Scope] ++ ZLayer.environment[ZIOAppArgs])
-      executionEventSinkLayer = sharedLayer
-      runner =
-        TestRunner(
-          TestExecutor
-            .default[
-              Environment,
-              Any
-            ](
-              sharedLayer,
-              perTestLayer,
-              executionEventSinkLayer,
-              eventHandlerZ
-            )
-        )
-      testReporter = testArgs.testRenderer.fold(runner.reporter)(createTestReporter)
-      summary <-
-        runner.withReporter(testReporter).run(aspects.foldLeft(filteredSpec)(_ @@ _) @@ TestAspect.fibers)
-    } yield summary
-  }
-
 }
