+++ test/shared/src/main/scala/zio/test/ZTestEventHandler.scala
@@ -1,7 +1,10 @@
 package zio.test
 
-import zio.UIO
+import zio.{UIO, ZIO}
 
 trait ZTestEventHandler {
-  def handle(event: ExecutionEvent.Test[_]): UIO[Unit]
+  def handle(event: ExecutionEvent): UIO[Unit]
+}
+object ZTestEventHandler {
+  val silent: ZTestEventHandler = _ => ZIO.unit
 }
