+++ test/shared/src/main/scala/zio/test/render/ExecutionResult.scala
@@ -17,7 +17,7 @@
 package zio.test.render
 
 import zio.stacktracer.TracingImplicits.disableAutoTrace
-import zio.test.TestAnnotationMap
+import zio.test.{SuiteId, TestAnnotationMap}
 import zio.test.render.ExecutionResult.Status._
 import zio.test.render.ExecutionResult.{ResultType, Status}
 import zio.test.render.LogLine.Line
@@ -35,6 +35,7 @@ case class ExecutionResult(
 
   def &&(that: ExecutionResult): ExecutionResult =
     (self.status, that.status) match {
+      case (Started, _)     => that
       case (Ignored, _)     => that
       case (_, Ignored)     => self
       case (Failed, Failed) => self.copy(streamingLines = self.streamingLines ++ that.streamingLines.tail)
@@ -44,6 +45,7 @@ case class ExecutionResult(
 
   def ||(that: ExecutionResult): ExecutionResult =
     (self.status, that.status) match {
+      case (Started, _)     => that
       case (Ignored, _)     => that
       case (_, Ignored)     => self
       case (Failed, Failed) => self.copy(streamingLines = self.streamingLines ++ that.streamingLines.tail)
@@ -53,6 +55,7 @@ case class ExecutionResult(
 
   def unary_! : ExecutionResult =
     self.status match {
+      case Started => self
       case Ignored => self
       case Failed  => self.copy(status = Passed)
       case Passed  => self.copy(status = Failed)
@@ -82,6 +85,7 @@ object ExecutionResult {
 
   sealed abstract class Status
   object Status {
+    case object Started extends Status
     case object Failed  extends Status
     case object Passed  extends Status
     case object Ignored extends Status
